backend/requirements.txt
google-adk>=1.26.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
redis>=5.2.0
livekit>=1.1.0
livekit-api>=1.0.0
websockets>=14.0
httpx>=0.28.0
structlog>=24.4.0
